#24 renamed dialogs files
