+++ src/qt_dialogs/dynamics_dialog.py
