+++ src/qt_dialogs/table_dialog.py
@@ -1,6 +1,6 @@
 from PyQt5.QtWidgets import QDialog
 
-from src.qt_models.pagemapModel import PagemapModel
+from src.qt_models.pagemap_model import PagemapModel
 from src.qt_ui.tableDialog_ui import Ui_TableDialog
